liquidations: bybit-метки long/short были перевёрнуты 2.5 месяца
Найдено при разборе /rynok: карточка написала «за 24ч вынесли шортов
84 vs 27», хотя цена за сутки упала — шорты выносит на РОСТЕ.

Причина: топик bybit allLiquidation отдаёт в поле S сторону ПОЗИЦИИ,
а коллектор трактовал её как сторону ордера (sell->long, правило
старого топика liquidation).

Доказательство на часовых ходах цены (лонги выносит на падении):
  bybit  май +0.132/-0.185  июнь +0.304/-0.345  июль +0.180/-0.179
         (метка long доминирует на РОСТЕ = инверсия, все 3 месяца)
  okx    май -0.197/+0.168  июнь -0.335/+0.327  июль -0.190/+0.174
         (корректно всё время)
bybit даёт ~2/3 объёма, поэтому портил агрегат целиком.

- фикс маппинга в коллекторе (sell->short);
- tools/fix_bybit_liq_sides.py: разовый переворот истории с бэкапом
  .pre-bybit-side-fix.bak (BTC 101.5k строк, ETH 61.8k, XRP 24.6k);
- после фикса обе биржи согласуются с ценой.

Затрагивает 19 потребителей поля side (cascade_alert, pre_cascade_alert,
scalp_liq, grid_coordinator, liquidity_map, watchlist и др.) — их выводы
на исторических данных строились на перевёрнутых метках.

+++ market_collector/liquidations.py
@@ -133,8 +133,15 @@ def _run_bybit_ws(stop_event: threading.Event) -> None:
                         for liq in liq_data:
                             if not isinstance(liq, dict):
                                 continue
+                            # 2026-07-23: у топика allLiquidation поле S — это
+                            # сторона ПОЗИЦИИ, а не ордера. Прежний маппинг
+                            # (sell→long, как для старого топика liquidation)
+                            # переворачивал метки. Проверено на 7д/150ч: с ним
+                            # метка «long» доминировала на РОСТЕ цены (+0.167%),
+                            # хотя лонги выносит на падении; у okx (корректный
+                            # маппинг) — ровно наоборот (−0.186%).
                             raw_side = liq.get("S") or liq.get("side", "")
-                            side = "long" if str(raw_side).lower() == "sell" else "short"
+                            side = "short" if str(raw_side).lower() == "sell" else "long"
                             qty = liq.get("v") or liq.get("size", "")
                             price = liq.get("p") or liq.get("price", "")
                             _write_liq(symbol, {
